Introducing a SavedObjectRepository (#19013)
* Create a separate SavedObjectRepository that the SavedObjectClient uses

* Moving the repository into lib/

* Fixing test after moving the repository

* Revising tests based on peer review

* Removing awaits

* Adding warning comments regarding the repository's impact on the SOC
