+++ src/server/saved_objects/client/lib/index.js
@@ -1,7 +1,4 @@
-export { getSearchDsl } from './search_dsl';
-export { trimIdPrefix } from './trim_id_prefix';
-export { includedFields } from './included_fields';
-export { decorateEsError } from './decorate_es_error';
+export { SavedObjectsRepository } from './repository';
 
 import * as errors from './errors';
 export { errors };
