+++ src/server/saved_objects/client/saved_objects_client.js
@@ -1,29 +1,11 @@
-import uuid from 'uuid';
-
-import { getRootType } from '../../mappings';
-
 import {
-  getSearchDsl,
-  trimIdPrefix,
-  includedFields,
-  decorateEsError,
+  SavedObjectsRepository,
   errors,
 } from './lib';
 
 export class SavedObjectsClient {
   constructor(options) {
-    const {
-      index,
-      mappings,
-      callCluster,
-      onBeforeWrite = () => {},
-    } = options;
-
-    this._index = index;
-    this._mappings = mappings;
-    this._type = getRootType(this._mappings);
-    this._onBeforeWrite = onBeforeWrite;
-    this._unwrappedCallCluster = callCluster;
+    this._repository = new SavedObjectsRepository(options);
   }
 
   /**
@@ -105,42 +87,7 @@ export class SavedObjectsClient {
    * @returns {promise} - { id, type, version, attributes }
   */
   async create(type, attributes = {}, options = {}) {
-    const {
-      id,
-      overwrite = false
-    } = options;
-
-    const method = id && !overwrite ? 'create' : 'index';
-    const time = this._getCurrentTime();
-
-    try {
-      const response = await this._writeToCluster(method, {
-        id: this._generateEsId(type, id),
-        type: this._type,
-        index: this._index,
-        refresh: 'wait_for',
-        body: {
-          type,
-          updated_at: time,
-          [type]: attributes
-        },
-      });
-
-      return {
-        id: trimIdPrefix(response._id, type),
-        type,
-        updated_at: time,
-        version: response._version,
-        attributes
-      };
-    } catch (error) {
-      if (errors.isNotFoundError(error)) {
-        // See "503s from missing index" above
-        throw errors.createEsAutoCreateIndexError();
-      }
-
-      throw error;
-    }
+    return this._repository.create(type, attributes, options);
   }
 
   /**
@@ -152,68 +99,7 @@ export class SavedObjectsClient {
    * @returns {promise} - [{ id, type, version, attributes, error: { message } }]
    */
   async bulkCreate(objects, options = {}) {
-    const {
-      overwrite = false
-    } = options;
-    const time = this._getCurrentTime();
-    const objectToBulkRequest = (object) => {
-      const method = object.id && !overwrite ? 'create' : 'index';
-
-      return [
-        {
-          [method]: {
-            _id: this._generateEsId(object.type, object.id),
-            _type: this._type,
-          }
-        },
-        {
-          type: object.type,
-          updated_at: time,
-          [object.type]: object.attributes
-        }
-      ];
-    };
-
-    const { items } = await this._writeToCluster('bulk', {
-      index: this._index,
-      refresh: 'wait_for',
-      body: objects.reduce((acc, object) => ([
-        ...acc,
-        ...objectToBulkRequest(object)
-      ]), []),
-    });
-
-    return items.map((response, i) => {
-      const {
-        error,
-        _id: responseId,
-        _version: version,
-      } = Object.values(response)[0];
-
-      const {
-        id = responseId,
-        type,
-        attributes,
-      } = objects[i];
-
-      if (error) {
-        return {
-          id,
-          type,
-          error: {
-            message: error.reason || JSON.stringify(error)
-          }
-        };
-      }
-
-      return {
-        id,
-        type,
-        updated_at: time,
-        version,
-        attributes
-      };
-    });
+    return this._repository.bulkCreate(objects, options);
   }
 
   /**
@@ -224,29 +110,7 @@ export class SavedObjectsClient {
    * @returns {promise}
    */
   async delete(type, id) {
-    const response = await this._writeToCluster('delete', {
-      id: this._generateEsId(type, id),
-      type: this._type,
-      index: this._index,
-      refresh: 'wait_for',
-      ignore: [404],
-    });
-
-    const deleted = response.result === 'deleted';
-    if (deleted) {
-      return {};
-    }
-
-    const docNotFound = response.result === 'not_found';
-    const indexNotFound = response.error && response.error.type === 'index_not_found_exception';
-    if (docNotFound || indexNotFound) {
-      // see "404s from missing index" above
-      throw errors.createGenericNotFoundError();
-    }
-
-    throw new Error(
-      `Unexpected Elasticsearch DELETE response: ${JSON.stringify({ type, id, response, })}`
-    );
+    return this._repository.delete(type, id);
   }
 
   /**
@@ -263,73 +127,7 @@ export class SavedObjectsClient {
    * @returns {promise} - { saved_objects: [{ id, type, version, attributes }], total, per_page, page }
    */
   async find(options = {}) {
-    const {
-      type,
-      search,
-      searchFields,
-      page = 1,
-      perPage = 20,
-      sortField,
-      sortOrder,
-      fields,
-      includeTypes,
-    } = options;
-
-    if (searchFields && !Array.isArray(searchFields)) {
-      throw new TypeError('options.searchFields must be an array');
-    }
-
-    if (fields && !Array.isArray(fields)) {
-      throw new TypeError('options.searchFields must be an array');
-    }
-
-    const esOptions = {
-      index: this._index,
-      size: perPage,
-      from: perPage * (page - 1),
-      _source: includedFields(type, fields),
-      ignore: [404],
-      body: {
-        version: true,
-        ...getSearchDsl(this._mappings, {
-          search,
-          searchFields,
-          type,
-          includeTypes,
-          sortField,
-          sortOrder
-        })
-      }
-    };
-
-    const response = await this._callCluster('search', esOptions);
-
-    if (response.status === 404) {
-      // 404 is only possible here if the index is missing, which
-      // we don't want to leak, see "404s from missing index" above
-      return {
-        page,
-        per_page: perPage,
-        total: 0,
-        saved_objects: []
-      };
-    }
-
-    return {
-      page,
-      per_page: perPage,
-      total: response.hits.total,
-      saved_objects: response.hits.hits.map(hit => {
-        const { type, updated_at: updatedAt } = hit._source;
-        return {
-          id: trimIdPrefix(hit._id, type),
-          type,
-          ...updatedAt && { updated_at: updatedAt },
-          version: hit._version,
-          attributes: hit._source[type],
-        };
-      }),
-    };
+    return this._repository.find(options);
   }
 
   /**
@@ -345,42 +143,7 @@ export class SavedObjectsClient {
    * ])
    */
   async bulkGet(objects = []) {
-    if (objects.length === 0) {
-      return { saved_objects: [] };
-    }
-
-    const response = await this._callCluster('mget', {
-      index: this._index,
-      body: {
-        docs: objects.map(object => ({
-          _id: this._generateEsId(object.type, object.id),
-          _type: this._type,
-        }))
-      }
-    });
-
-    return {
-      saved_objects: response.docs.map((doc, i) => {
-        const { id, type } = objects[i];
-
-        if (!doc.found) {
-          return {
-            id,
-            type,
-            error: { statusCode: 404, message: 'Not found' }
-          };
-        }
-
-        const time = doc._source.updated_at;
-        return {
-          id,
-          type,
-          ...time && { updated_at: time },
-          version: doc._version,
-          attributes: doc._source[type]
-        };
-      })
-    };
+    return this._repository.bulkGet(objects);
   }
 
   /**
@@ -391,29 +154,7 @@ export class SavedObjectsClient {
    * @returns {promise} - { id, type, version, attributes }
    */
   async get(type, id) {
-    const response = await this._callCluster('get', {
-      id: this._generateEsId(type, id),
-      type: this._type,
-      index: this._index,
-      ignore: [404]
-    });
-
-    const docNotFound = response.found === false;
-    const indexNotFound = response.status === 404;
-    if (docNotFound || indexNotFound) {
-      // see "404s from missing index" above
-      throw errors.createGenericNotFoundError();
-    }
-
-    const { updated_at: updatedAt } = response._source;
-
-    return {
-      id,
-      type,
-      ...updatedAt && { updated_at: updatedAt },
-      version: response._version,
-      attributes: response._source[type]
-    };
+    return this._repository.get(type, id);
   }
 
   /**
@@ -426,58 +167,6 @@ export class SavedObjectsClient {
    * @returns {promise}
    */
   async update(type, id, attributes, options = {}) {
-    const time = this._getCurrentTime();
-    const response = await this._writeToCluster('update', {
-      id: this._generateEsId(type, id),
-      type: this._type,
-      index: this._index,
-      version: options.version,
-      refresh: 'wait_for',
-      ignore: [404],
-      body: {
-        doc: {
-          updated_at: time,
-          [type]: attributes
-        }
-      },
-    });
-
-    if (response.status === 404) {
-      // see "404s from missing index" above
-      throw errors.createGenericNotFoundError();
-    }
-
-    return {
-      id,
-      type,
-      updated_at: time,
-      version: response._version,
-      attributes
-    };
-  }
-
-  async _writeToCluster(method, params) {
-    try {
-      await this._onBeforeWrite();
-      return await this._callCluster(method, params);
-    } catch (err) {
-      throw decorateEsError(err);
-    }
-  }
-
-  async _callCluster(method, params) {
-    try {
-      return await this._unwrappedCallCluster(method, params);
-    } catch (err) {
-      throw decorateEsError(err);
-    }
-  }
-
-  _generateEsId(type, id) {
-    return `${type}:${id || uuid.v1()}`;
-  }
-
-  _getCurrentTime() {
-    return new Date().toISOString();
+    return this._repository.update(type, id, attributes, options);
   }
 }
